refactor: 예약 및 접수 관련 공용 컴포넌트들 /src/common/components로 이동.
