+++ frontend/src/reception/views/WaitingList.vue
@@ -1,8 +1,8 @@
 <script setup>
   import {useWaitingListStore} from "@/stores/waitingListStore.js";
   import {onBeforeMount, onMounted, onUnmounted, ref} from "vue";
-  import WaitingListDoctorName from "@/reception/components/WaitingListDoctorName.vue";
-  import WaitingListPatientList from "@/reception/components/WaitingListPatientList.vue";
+  import WaitingListDoctorName from "@/common/components/WaitingListDoctorName.vue";
+  import WaitingListPatientList from "@/common/components/WaitingListPatientList.vue";
   import {reception} from "@/reception/util/reception.js";
 
   const waitingListStore = useWaitingListStore();
